tchannel: re-organize functions for better discoverability
This minor change moves shared logic from the generally unsued
`ChannelOutbound` to the traditional `Outbound` and an error handling
method to `errors.go`.

No functionality has changed.
