+++ transport/tchannel/outbound.go
@@ -22,12 +22,14 @@ package tchannel
 
 import (
 	"context"
+	"io"
 	"strconv"
 
 	"github.com/uber/tchannel-go"
 	"go.uber.org/yarpc/api/peer"
 	"go.uber.org/yarpc/api/transport"
 	"go.uber.org/yarpc/api/x/introspection"
+	"go.uber.org/yarpc/internal/iopool"
 	intyarpcerrors "go.uber.org/yarpc/internal/yarpcerrors"
 	peerchooser "go.uber.org/yarpc/peer"
 	"go.uber.org/yarpc/peer/hostport"
@@ -194,6 +196,36 @@ func callWithPeer(ctx context.Context, req *transport.Request, peer *tchannel.Pe
 		},
 	}
 	return resp, err
+}
+
+func writeBody(body io.Reader, call *tchannel.OutboundCall) error {
+	w, err := call.Arg3Writer()
+	if err != nil {
+		return err
+	}
+
+	if _, err := iopool.Copy(w, body); err != nil {
+		return err
+	}
+
+	return w.Close()
+}
+
+func getResponseError(headers transport.Headers) error {
+	errorCodeString, ok := headers.Get(ErrorCodeHeaderKey)
+	if !ok {
+		return nil
+	}
+	var errorCode yarpcerrors.Code
+	if err := errorCode.UnmarshalText([]byte(errorCodeString)); err != nil {
+		return err
+	}
+	if errorCode == yarpcerrors.CodeOK {
+		return yarpcerrors.Newf(yarpcerrors.CodeInternal, "got CodeOK from error header")
+	}
+	errorName, _ := headers.Get(ErrorNameHeaderKey)
+	errorMessage, _ := headers.Get(ErrorMessageHeaderKey)
+	return intyarpcerrors.NewWithNamef(errorCode, errorName, errorMessage)
 }
 
 func getApplicationErrorCodeFromHeaders(headers transport.Headers) *yarpcerrors.Code {
